Move extensions into sub-directories
