+++ bot/cogs/backend/error_handler.py
