+++ bot/cogs/filters/antimalware.py
