+++ bot/cogs/filters/filter_lists.py
