+++ bot/cogs/filters/security.py
