+++ bot/cogs/filters/token_remover.py
