+++ bot/cogs/filters/webhook_remover.py
