+++ bot/cogs/info/doc.py
