+++ bot/cogs/info/information.py
