+++ bot/cogs/info/tags.py
