+++ bot/cogs/moderation/infraction/superstarify.py
