+++ bot/cogs/moderation/watchchannels/talentpool.py
