+++ bot/cogs/utils/clean.py
