+++ bot/cogs/utils/jams.py
