+++ tests/bot/cogs/sync/test_roles.py
@@ -3,7 +3,7 @@ from unittest import mock
 
 import discord
 
-from bot.cogs.sync.syncers import RoleSyncer, _Diff, _Role
+from bot.cogs.backend.sync import RoleSyncer, _Diff, _Role
 from tests import helpers
