+++ tests/bot/cogs/test_antimalware.py
@@ -3,7 +3,7 @@ from unittest.mock import AsyncMock, Mock
 
 from discord import NotFound
 
-from bot.cogs import antimalware
+from bot.cogs.filters import antimalware
 from bot.constants import Channels, STAFF_ROLES
 from tests.helpers import MockAttachment, MockBot, MockMessage, MockRole
