+++ tests/bot/cogs/test_token_remover.py
@@ -6,9 +6,9 @@ from unittest.mock import MagicMock
 from discord import Colour, NotFound
 
 from bot import constants
-from bot.cogs import token_remover
+from bot.cogs.filters import token_remover
+from bot.cogs.filters.token_remover import Token, TokenRemover
 from bot.cogs.moderation import ModLog
-from bot.cogs.token_remover import Token, TokenRemover
 from tests.helpers import MockBot, MockMessage, autospec
